pyproject.toml
[tool.poetry]
name = "fastapi-realworld-example-app"
version = "0.1.0"
description = "Backend logic implementation for https://github.com/gothinkster/realworld with awesome FastAPI"
authors = ["qiucheng <qiucheng@jinniuai.com"]
license = "MIT"


[[tool.poetry.source]]
name = 'tsinghua'
url = 'https://pypi.tuna.tsinghua.edu.cn/simple'

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.6.1"
gunicorn = "^19.9"
fastapi = "0.61.1"
pydantic = {version = "1.6.1",extras = ["email"]}
passlib = {version = "^1.7",extras = ["bcrypt"]}
pyjwt = "^1.7"
alembic = "^1.0"
python-slugify = "^3.0"
Unidecode = "^1.0"
python-dotenv = "^0.10.1"
databases = "^0.2.1"
motor = "^2.0"
python-multipart = "^0.0.5"
aiofiles = "^0.5.0"


[tool.poetry.dev-dependencies]
pytest = "*"
autopep8 = "^1.5.4"

[tool.black]
exclued = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

